pyproject.toml
[project]
name = "claimr"
version = "0.1.0"
description = "Claim Account Manager"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dataset>=1.6.2",
    "flask>=3.1.0",
    "hive-nectar>=0.0.5",
    "python-dotenv>=1.1.0",
]
